move ui component under function folder
