+++ packages/app/src/cli/services/function/replay.ts
@@ -3,7 +3,7 @@ import {AppInterface} from '../../models/app/app.js'
 import {ExtensionInstance} from '../../models/extensions/extension-instance.js'
 import {FunctionConfigType} from '../../models/extensions/specifications/function.js'
 import {selectFunctionRunPrompt} from '../../prompts/function/replay.js'
-import {renderReplay} from '../dev/ui.js'
+import {renderReplay} from './ui.js'
 
 import {joinPath} from '@shopify/cli-kit/node/path'
 import {readFile} from '@shopify/cli-kit/node/fs'
